pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openwater-thermal-analyst"
version = "0.1.0"
description = "Integrates thermal infrared remote sensing, hydrodynamic modeling, compliance monitoring, ML prediction, and alerting for water thermal plume analysis"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "OpenWaterThermalAnalyst Authors" }]
license = { text = "MIT" }
keywords = ["thermal", "remote sensing", "hydrodynamics", "water quality", "compliance", "plume", "Landsat", "MODIS", "ECOSTRESS"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Intended Audience :: Science/Research",
  "Topic :: Scientific/Engineering :: GIS",
  "Topic :: Scientific/Engineering :: Hydrology",
]

# Core dependencies kept moderate; some heavy libs are optional extras.
dependencies = [
  "numpy>=1.24",
  "scipy>=1.11",
  "pandas>=2.0",
  "scikit-learn>=1.3",
  "requests>=2.31",
  "PyYAML>=6.0",
  "xarray>=2023.1",
  "netCDF4>=1.6",
  "rasterio>=1.3",
  "shapely>=2.0",
  "fiona>=1.9",
  "matplotlib>=3.8",
]

[project.optional-dependencies]
ml = [
  "torch>=2.2",
]
hydro = [
  "meshio>=5.3",
]

[project.scripts]
openwater = "open_water_thermal_analyst.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["open_water_thermal_analy*", "open_water_thermal_analyst*"]

